updated type defintion, removed duplicate code and added mockData to integrationTestData
