+++ server/tests/allocation.e2e.test.js
@@ -5,32 +5,7 @@ const request = require('supertest');
 const app = require('../server');
 const { startDB, closeDB } = require('./util/db');
 const { getKeycloakToken, ministryOfHealth, healthAuthority } = require('./util/keycloak');
-const { createGlobalPhase } = require('../services/phase');
-const { makeTestSite } = require('./util/integrationTestData');
-
-const dataSetup = async (id) => {
-  const site = await makeTestSite({
-    siteId: id,
-    siteName: 'Test Site 1040',
-    city: 'Test City 1040',
-  });
-
-  const phaseData = {
-    name: 'Test Phase',
-    start_date: new Date(),
-    end_date: new Date(),
-  };
-  const user = {
-    id: 'noid',
-  };
-  expect(site.siteId).toBeDefined();
-  const phase = await createGlobalPhase(phaseData, user);
-  expect(phase.id).toBeDefined();
-  return {
-    site,
-    phase,
-  };
-};
+const { makeTestFKAllocations } = require('./util/integrationTestData');
 
 describe('api e2e tests for /allocation', () => {
   let server;
@@ -45,7 +20,7 @@ describe('api e2e tests for /allocation', () => {
   });
 
   it('should set allocation to phase', async () => {
-    const { site, phase } = await dataSetup(52345);
+    const { site, phase } = await makeTestFKAllocations(52345);
     const header = await getKeycloakToken(ministryOfHealth);
     const res = await request(app)
       .post(`/api/v1/allocation`)
@@ -116,7 +91,7 @@ describe('api e2e tests for /allocation', () => {
   });
 
   it('should fail to set an allocation due to unauthorized user', async () => {
-    const { site, phase } = await dataSetup(23621346);
+    const { site, phase } = await makeTestFKAllocations(23621346);
     const header = await getKeycloakToken(healthAuthority);
     const res = await request(app)
       .post(`/api/v1/allocation`)
